pyproject.toml
[tool.poetry]
name = "Virtual"
version = "0.1.0"
description = "Virtual uses hand tracking to emulate common input devices like the mouse, keyboard and volume control systems."
authors = ["Souhit Dey <obitogod@protonmail.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
mediapipe = "^0.8.8"
pyalsaaudio = {version = "^0.9.0", platform = "linux"}
pycaw = {version = "^20181226", platform = "win32"}
opencv-python = "4.5.3.56"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
mypy = "^0.910"
pytest = "^6.2.5"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
